merge

+++ src/zork/Game.java
@@ -88,7 +88,7 @@ public class Game {
       String itemId = (String) ((JSONObject) itemObj).get("id");
       String itemDescription = (String) ((JSONObject) itemObj).get("description");
       String itemRoom = (String) ((JSONObject) itemObj).get("room");
-      int itemWeight = (int) ((JSONObject) itemObj).get("weight");
+      int itemWeight = (int)((long)((JSONObject) itemObj).get("weight"));
       boolean itemCanEat = (boolean) ((JSONObject) itemObj).get("canEat");
       boolean itemIsTask = (boolean) ((JSONObject) itemObj).get("isTask");
       item = new Item(itemWeight, itemName,false, itemCanEat, itemIsTask, itemRoom, itemDescription);
